Handle null MIME type
Ideally, there is always a mimetype provided when attempting to import a HAR file.
However in practice when importing HAR files from proxy sources (such as Charles),
you may receive requests with no MIME type. Currently this fails the import for
the entire HAR.

This change will cause HAR imports to succeed even when a HAR has a request with
a missing MIME type. I decided to use this behavior because it roughly mimics what
Mozilla does.

https://developer.mozilla.org/en-US/docs/Mozilla/How_Mozilla_determines_MIME_Types#Unknown_Decoder